Put transaction.atomic() around creation of TestModel instance
See https://docs.djangoproject.com/en/1.9/topics/db/transactions/#controlling-transactions-explicitly. We need to make sure the query that raised the
DatabaseError has been rolled back before handling the exception;
otherwise we'll get a TransactionManagementError next time we try to do a
db read or write. We were seeing this when trying to get the result of the
celery health check after the db health check had failed with a
DatabaseError, because the celery check tries to read the task result from
the database.